[IMP] product_packaging_volume

+++ product_packaging_volume/models/stock_move_line.py
@@ -9,11 +9,11 @@ class StockMoveLine(models.Model):
     packaging_id = fields.Many2one(
         string='Package Type', comodel_name='product.packaging',
         related='result_package_id.packaging_id', store=True)
-    height = fields.Integer(
+    height = fields.Float(
         string='Height', related='packaging_id.height', store=True)
-    width = fields.Integer(
+    width = fields.Float(
         string='Width', related='packaging_id.width', store=True)
-    packaging_length = fields.Integer(
+    packaging_length = fields.Float(
         string='Packaging Length', related='packaging_id.packaging_length',
         store=True)
     volume = fields.Float(
